requirements.txt
# Django & DRF
Django==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Database (SQLite comes with Python, but add PostgreSQL for production)
# psycopg2-binary==2.9.9  # Uncomment for production

# Image processing
Pillow==10.2.0

# Django filters
django-filter==23.5

# Environment variables (for production)
python-dotenv==1.0.1
